requirements.txt
Flask==2.3.2
gunicorn==20.1.0
aiohttp==3.8.4
pymongo
pymongo[srv]==3.12.3
python-telegram-bot==21.2
python-telegram-bot[job-queue]
Pillow
requests
pytz
